Add NOT NULL constraints to fixes table and consolidate Fix types
- Add database migration for NOT NULL constraints on via, device_address, address_type, device_id
- Consolidate position::Fix and fixes::Fix into single fixes::Fix struct
- Add Diesel derive attributes (Queryable, Selectable) to fixes::Fix
- Update device_address to use hex string representation with device_address_hex() method
- Remove redundant position.rs module
