+++ src/schema.rs
@@ -246,7 +246,7 @@ diesel::table! {
         source -> Varchar,
         #[max_length = 9]
         destination -> Varchar,
-        via -> Nullable<Array<Nullable<Text>>>,
+        via -> Array<Nullable<Text>>,
         raw_packet -> Text,
         timestamp -> Timestamptz,
         latitude -> Float8,
@@ -254,8 +254,8 @@ diesel::table! {
         location -> Nullable<Geography>,
         altitude_feet -> Nullable<Int4>,
         #[max_length = 10]
-        device_address -> Nullable<Varchar>,
-        address_type -> Nullable<AddressType>,
+        device_address -> Varchar,
+        address_type -> AddressType,
         aircraft_type_ogn -> Nullable<AircraftTypeOgn>,
         #[max_length = 20]
         flight_number -> Nullable<Varchar>,
@@ -276,7 +276,7 @@ diesel::table! {
         club_id -> Nullable<Uuid>,
         flight_id -> Nullable<Uuid>,
         unparsed_data -> Nullable<Varchar>,
-        device_id -> Nullable<Uuid>,
+        device_id -> Uuid,
         received_at -> Timestamptz,
         lag -> Nullable<Int4>,
     }
